Fixed projectile paths + Span writer/reader lol

+++ AlloyClient/Game/Objects/ProjectilePaths/AcceleratePath.cs
@@ -1,7 +1,7 @@
 ﻿#region
 
 using System;
-using System.Numerics;
+using OpenTK.Mathematics;
 
 #endregion
 
@@ -15,7 +15,7 @@ public class AcceleratePath : ProjectilePathSegment
         : base(PathType.AcceleratePath, speed, angle, lifetimeMs, timeOffset, mods)
     { }
 
-    public override Vector2 PositionAt(int elapsedLifetimeMs)
+    public override Vector2 PositionAt(float elapsedLifetimeMs)
     {
         var speed = Speed;
         var p = Vector2.Zero;
